added Beginners Planting Tips function

+++ lib/src/screens/logged_in_screens/home_screen/home_screen.dart
@@ -113,7 +113,7 @@ class _HomeScreenState extends State<HomeScreen> {
   Widget _buildCheckTempBtn(BuildContext context) {
     return TextButton(
       onPressed: () {
-        if (int.parse(temperature) >= 30) {
+        if (double.parse(temperature) >= 30) {
           TopSnackBar.showError(
             context: context,
             errorMessage:
